Allow stdlib to be documented with docstrings
Update the func_def!() macro to require doc-strings describing the
function and store them in the FuncDef structure where the are
accessible at runtime.

Add an --output-docs parameter to resynth which allows the stdlib
documentation to be written out to a markdown tree structure. The
generated markdown, and the quality of the documentation right now is..
well.. poor. But we can improve it later.

Also, methods on classes are not found by the documentation walk because
they aren't linked in to the STDLIB module tree anywhere. We will need
to fix that somehow.

But this is just to prove the basic mechanism that resynth can
introspect and generate its own docs.