Add auto-generated documentation
Improve the formatting of the documentation, and include the generated
documentation in the repo so that it can be browsed on github.

Some of the documentation is pretty sparse, but at least spruce up the
TLS documentation by including an example and documenting all the basic
functions. Let's just see what this puppy can do :)